// apps/web-svelte/src/lib/api/generated/model/entityCapabilityDescriptionCapability.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { EntityCapabilityDescriptionCapabilityKind } from './entityCapabilityDescriptionCapabilityKind';

export interface EntityCapabilityDescriptionCapability {
  kind: EntityCapabilityDescriptionCapabilityKind;
  value: string;
}