// apps/web-svelte/src/lib/api/generated/model/entityCapabilityRatingCapability.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { EntityCapabilityRatingCapabilityKind } from './entityCapabilityRatingCapabilityKind';

export interface EntityCapabilityRatingCapability {
  kind: EntityCapabilityRatingCapabilityKind;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  value: number | string | null;
}