// apps/web-svelte/src/lib/api/generated/model/entityCapabilityTechnicalCapability.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { EntityCapabilityTechnicalCapabilityKind } from './entityCapabilityTechnicalCapabilityKind';

export interface EntityCapabilityTechnicalCapability {
  kind: EntityCapabilityTechnicalCapabilityKind;
  /**
     * @nullable
     * @pattern ^-?(\d+\.)?\d{2}:\d{2}:\d{2}(\.\d{1,7})?$
     */
  duration: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  width: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  height: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
     */
  frameRate: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  bitRate: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  sampleRate: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  channels: number | string | null;
  /** @nullable */
  codec: string | null;
  /** @nullable */
  container: string | null;
  /** @nullable */
  format: string | null;
}